requirements.txt
d20==1.1.2
discord.py==2.4.0
python-dotenv==1.0.1
pillow==11.1.0